[LT] Compose DDP with LazyTensor
Summary:
This patch adds the support of DDP to LazyTensor. It's done in two manners:
1) Defines a LazyProcessGroup which is a subclass of dist.ProcessGroup that
overrides comm ops: allgather, allreduce, and broadcast which are needed for
DDP. What they do is basically converting any lazy tensors into cuda tensors
and then replay the comm op in nccl. This is so called fallback approach.
2) Provides an additional lazy_comm_hook for overriding the all_reduce op
in DDP. This is a tentative workaround since we can't set the Future object
in dist._Work for the first approach, which the Reducer relies on to get
the finalized gradients. To be noted, other comm ops work because the callers
just reuse the passed in tensors as the results.
3) Please refer to the test case for how to init LazyProcessGroup and register
lazy_comm_hook.

Test Plan:
python test/lazy/test_ddp.py

ghstack-source-id: 2b8bba6fc76f4a5690d34b50a09b2fd2af8693f9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75350